break out node specific config logic
