+++ packages/cli-lib/package.json
@@ -9,7 +9,7 @@
     "url": "https://github.com/HubSpot/hubspot-cms-tools"
   },
   "dependencies": {
-    "chalk": "^2.4.2",
+    "chalk": "^4",
     "chokidar": "^3.0.1",
     "content-disposition": "^0.5.3",
     "debounce": "^1.2.0",
